update password + login pages
